Fix wordCheck, Edit searcher
The logic of wordCheck was overhauled.
The current logic seems to be correct so far,
even increasing the possibility of words being added.

seacher was edited for efficiency
Words are removed from a list after they are found
since duplicates are not allowed by buildGrid.
Similiarly, a position is not searched in after
a word has already been found there.

More statements added for verbose mode of searcher