Modify searcher, one place many words
A single location in the grid could hold
the starting letter to multiple words (up to 8).
Hence searcher was modified to keep looking
for words possibilities at one location
even after a word has already been found.

SPREAD and SPLIT words added in the
test case of run.py to give an example.