Improve searcher
The logic of searcher checked whether
any position held the first letter of a word to be found.
After this it was meaningless to check for
words which did not start with that letter.
Hence, f_word was made a dictionary that
stored each word by its first letter.
When a word is found, it is knocked out of the
dictionary so that it may not be looked for again.
To allow finding multiple words from
the same starting point,
w_f was turned into a list, instead of a string
so that multiple words could be knocked off
from f_word at once.